refactor: Moved the is logged in function to App component
