+++ src/Components/NavBar/NavBar.js
@@ -3,11 +3,10 @@ import './NavBar.scss';
 import { NavLink, useNavigate, useLocation } from 'react-router-dom';
 import logoImage from '../../images/refuge.png';
 
-export default function NavBar({ isLoggedIn, setIsLoggedIn }) {
+export default function NavBar({ isLoggedIn, handleLogin }) {
 
 
   console.log("NavBar rendering. isLoggedIn:", isLoggedIn);
-  console.log(typeof setIsLoggedIn); //dont forget to delete
 
   const navigate = useNavigate();
   const location = useLocation();
@@ -15,17 +14,6 @@ export default function NavBar({ isLoggedIn, setIsLoggedIn }) {
   useEffect(() => {
     console.log("NavBar Mounted"); //dont forget to delete
   }, []);
-
-  const handleOrganizationLogin = () => {
-    if (isLoggedIn) {
-      setIsLoggedIn(false); 
-      console.log("Attempting to logout and redirect to homepage");
-      navigate('/');
-    } else {
-      setIsLoggedIn(true);
-      navigate('/organizationDashboard');
-    }
-  }
 
   return (
     <nav className='nav-bar'>
@@ -43,15 +31,16 @@ export default function NavBar({ isLoggedIn, setIsLoggedIn }) {
         {isLoggedIn ? (
           <>
             <NavLink className='nav-link' to='/organizationDashboard'>organization dashboard</NavLink>
-            {/* <button className='nav-button logout' onClick={() => { setIsLoggedIn(false); }}>logout</button> */}
-            <button className='nav-button logout' onClick={() => { console.log('Logout button clicked!'); setIsLoggedIn(false); }}>logout</button>
-
+            <button className='nav-button logout' onClick={handleLogin}>logout</button>
 
           </>
         ) : (
-          <button className='nav-button' onClick={handleOrganizationLogin}>organization login</button>
+          <button className='nav-button' onClick={handleLogin}>organization login</button>
         )}
       </div>
     </nav>
   );
 }
+
+
+
